requirements.txt
torch==1.6.0
torchvision==0.7.0
numpy==1.18.1
streamlit==0.65.0
Pillow==8.0.1
